Create the Add expression
https://github.com/makingthematrix/replcalc/projects/1#card-74101796

Add and Substract have the same priority which in practice means they will be the same expression type -
the only difference will be in the actual evaluation of the term.

This commit introduces only Add. In contrast to the old calculator Add this time I don't try to be too
smart and every Add will have only two sub-expressions: left and right. So if we have "1 + 2 + 3",
it will be parsed into Add(Add(1, 2), 3). Easier for recursion. There will a lot of recursion in this
project.
By the way, the order of nesting is actually important here. Add is associative so it wouldn't matter
if we had only Add, but Substract is not and it relies on that we go from left to right. For exmaple,
"1 + 2 + 3" gives the same result no matter which two numbers we will add first,  but "3 - 2 - 1"
will give us 0 if we first substract 2 from 3 and then 1 from the result of "3 - 2", and it will
give us 2 fi we mess up and first substract 1 from 2.

I also made some refactoring that should make it easier in the future to introduce new expression types.