feat: Create the Multiply/Divide expression
https://github.com/makingthematrix/replcalc/projects/1#card-74101803

I decided to write this together in one commit because it's basically a logical copy of the previous chunk.
The multiply/divide expression is parsed in the same way as the add/substract, we just look for different operators.

One curious thing you may notice is that we say that addition and substraction has *lower* priority than
multiplication and division when it comes to the *evaluation* of a complex expression, but while *parsing*
this is reverted. We first need to look for "+" and "-", split the expression along these lines, and only then
look for "*" and "/" in the chunks we got. This way we will create our Abstract Syntax Tree. And  then, during *evaluation*,
we will go down the tree and start to evaluate it from bottom-up, so *first* we will evaluate multiplication and
division, and only then addition and substraction, which we will find higher in the tree.