feat: Unary minus
https://github.com/makingthematrix/replcalc/projects/1#card-74101815

The simplest way to implement a unary "-" operator is to skip it while looking for binary "-".
For now the only other place where we can find a unary "-" is in front of numbers, and that
means they can be a part of the Constant expression.

To identify and skip them, I changed the way AddSubstract looks for the "-" operator.
If there is "-" in the text, the program looks what is in front of it. If it's one of the operators
or the "-" operator is at the very beginning of the text, then it's a unary "-" and we need to
keep looking, moving one character back, from the end of the text to the front.
Otherwise it's a binary one and this is what we look for.

This will however get more complicated with the introduction of parentheses, so in anticipation
of that I will do some refactoring in the next commit.