Unary "-" as an expression
This is just refactoring in anticipation of that I will have to manage the unary "-" not only in front of constants
(where it is handled automatically) but also before parentheses, values, and functions. They will be modelled
as expressions as well, so the unary "-" needs to be also an expression with one inner expression in it.