The main loop
https://github.com/makingthematrix/replcalc/projects/1#card-74101894

This is just a very simple main loop that reads lines from the input, try to evaluate them and display them,
or ends if the input line is ":exit".

I have an idea for the future that lines starting with ":" will be commands instead of expressions.
Also, I needed to remove `fork := true` from `build.sbt` because otherwise reading from the input ends at once.
And I used this opportunity to make the code a bit more Scala-3-like - I removed braces where I could.