feat: Display meaningful errors #1 - from NaN to Option
https://github.com/makingthematrix/replcalc/projects/1#card-74101918

Until now I used `Double.NaN` to indicate that there was an error during parsing or evaluating.
In the end, I'd like the REPL to display a meaningful error about what was the problem.
There are a few ways to do it and at this point I haven't decided yet, but the first step is already clear:
I'm changing the result of evaluation to Option[Double] and so the incorrect evaluation will be
represented by `None` instead of `Double.NaN`. If the program fails to evaluate one of the inner expressions,
it generates `None` and it should bubble up to the top.
In the next commits I will move from `Option[Double]` to a type that is able to also describe the error
a bit better. Maybe `Either`, maybe an enum. This refactoring will already make it a bit easier.