feat: Display meaningful errors #2 - refactoring
Here I try to make the `Expression` class a bit more generic so that when I change the result types of parsing and
evaluating it won't require so big changes. Instead of relying on `Option.orElse` for traversing possible stages
(e.g. should I now parse it as a "+"? No? A "*" then? Also no? How about a unary "-"?...) I put those stages in
a sequence of functions and then use `Seq.collectFirst` to find the first that can handle the expression.