feat: Display meaningful errors #3 - From Option[Expression] to Option[Either[Error, Expression]]
The Left of Either carries errors, the None of Option means we should look further in the sequence of stages
as the current one is not applicable. Option[Either[Error, Expression]] is a bit too wordy for my liking
(so I created an alias), but the alternative I considered, an enum with cases for a valid expression, an error,
and a "none", had its own issues too.