feat: Write an expression type for assignment
https://github.com/makingthematrix/replcalc/projects/1#card-74102015

I rewrote the old Text class as Assignment.
Before I thought I could use Text for storing chunks of the expressions that I would like to parse lazily -
expressions in parentheses, and maybe values. But since then I split the processing into two, parsing and
evaluating, and the Text class became useless.

The new expression type, Assignment, stores an expression in the parsed, but not yet evaluated form,
under a given name. The name needs to start with a letter or an underscore and can be followed by letters,
digits, and the underscore. For now it doesn't do much more than that - it's just an expression with a name.
In the future though, when it turns out that the line of text is parsed as an assignment, it will be
stored in memory for further use. There will be another expression type, Value, that will wrap around
a name used in any more complex expression, and when it's evaluated, it will search for an assignment with
the same name.