feat: Introduce a dictionary for assigned expressions
https://github.com/makingthematrix/replcalc/projects/1#card-74102015

Assignments have side-effects: they not only parse expressions, but also save them in the memory for further use.
For a moment I experimented with an idea how to do it and still avoid side-effects. It led me to write something
like a Finite State Machine where the whole REPL is na FSM, it starts with some initial state, with each line
given by the user the machine generates a new state, and then awaits for another line within this new state.
But I decided that it would be too complicated for what I want to do here. Instead, I made a mutable `Dictionary`
which is passed around from one expression to another and slowly accumulates named expressions. It's not as
"pure FP" as it could be but it works well. It's an example of how everything in programming can be done in more
than one way.

Mind you, it's still more FP than if I made the `Dictionary` as a singleton and then every `Assignment` would have
a hardcoded reference to it. That would be ugly. And not good for unit testing.