feat: List values in REPL
Here I introduce a second command, after ":exit". This one is called ":list" and displays all the values that were
accumulated in the dictionary so far. Note that if there was an error during parsing of the assignment, the error
will prevent the assignment from being added to the dictionary. But, for example, "a = 1 / 0" is a perfectly valid
assignment. It will be added and only during evaluation we will get an error that we try to divide by zero.
That's why it is possible that we will see errors on the displayed list.