feat: Use values in expressions
https://github.com/makingthematrix/replcalc/projects/1#card-74102091

After a valid assignment, the user is now able to use the name of the assigned expression as a value in all
later expressions. So, if for example I write first "a = 1", from this moment I'm able to do something like
"b = a + 1". The expression assigned to "a" will be taken from the dictionary and will become a part of a more
complex expression that will be assigned to "b", which from this moement I can also use to create even more
complex expressions.

Maybe this is also a good moment to answer the questions what about reassignments - if I have "a = 1",
can I now do "a = 2"? And the answer is "no". Choose another name. The reason is that this is supposed to be
just a calculator, not a compiler of a small programming language (yet) so I think it's more important for
the user to be safe and unable to make errors than to be able to introduce more complexity. With reassignments
and the fact that the program evaluates expressions on demand (expressions are "lazy", so to say), this would
be possible:
```
a = 0
b = a + 1
a = b + 1
a
```
Here, first we create "a", then "b", and then we reassign "a" so that now "b" is an inner expression of "a"
and "a" is an inner expression of "b". And then we try to evaluate "a" which starts an infinite recursion of
evaluation.

So, no. No reassignments. At least for now.