feat: Create the preprocessor class
https://github.com/makingthematrix/replcalc/projects/1#card-74102386

While the parser transforms a line of text into an AST of expressions, the preprocessor only performs
some operation on the line, making it easier for the parser to work on it later.
The first task of the preprocessor will be to remove all the whitespaces - they are never significant
and their presence in the line make the parsing more complicated (all the trimming...).