refactor: Refactor Parser for more flexible use
So far the `Parser` functionality was simple enough that everything could be contained within one method, `parse`.
But after I made `Dictionary`, and now with `Preprocessor`, the code starts to get more complex.
To make the future work easier I decided to change `Parser` from a Scala `object` into a regular `class` instance
with references to the preprocessor and the dictionary.