feat: Handle parentheses
https://github.com/makingthematrix/replcalc/projects/1#card-74101822

There is no special expression type for parentheses. Instead, parentheses are turned to values by the preprocessor.
Each pair of parentheses is parsed as an individual, smaller expression, added to the dictionary under a special name,
and then it's replaced in the original like by that special name. For example "1 + (2 + 3) + 4" becomes "1 + $1 + 4"
where "$1" is the name of the value. In the dictionary we now have an expression "2 + 3" under this name. When
evaluated, the value "$1" will be replaced with the result of "2 + 3" and used to evaluate the rest of the higher-level
expression.