refactor: Find expressions for values at evaluation
Until now when I found a name of a value in the parsed line, I pulled the associated expression from the dictionary
during the parsing stage and made a reference to it in the value. That works well for values - which are constant -
but when thinking how to implement functions I realized that for functions I won't have expressions ready at the
parsing stage. Functions have arguments and expressions associated with arguments are not known when a function is
created (parsing) but only when it is used (evaluation). During the parsing I prepare a template, so to say, where
certain parts are waiting to be filled later.

I decided that `Value` will be used for such holes in the template. From now on a value will only hold its name,
so it will be possible to use it in the template, as the program won't search for the associated expression that
doesn't exist yet. Only when the function is evaluated, the program will first parse the arguments and prepare
required expressions, and then when it will find a value during the evaluation of the function, it will know
what expression it refers to.