feat: Implement assignments to functions
https://github.com/makingthematrix/replcalc/projects/1#card-74102362

This PR allows to create a new type of assignment, a function assignment, and add it to the dictionary.
It's not possible to use it yet.
The left side of the assignment needs to be a name of the function followed by the opening parenthesis,
a list of arguments names spearated with commas, and the closing parenthesis.
The right side can be any expression that uses previously assigned values and arguments defined on the left side.
In theory it should also possible to use other functions, but for now evaluation of functions is not implemented, so no.