feat: Enable copying and updating dictionaries
In preparation for inplementing functions, I want first to get a few things out of the way.
One of them is that I will need more than one global storage for expressions, i.e. more than one dictionary.
Let's say we have global values `a` and `b`, and then we create a function `foo(x)`.
The global dictionary should then hold expressions for `a`, `b`, and `foo`, but not for `x`.
`x` will only exist within the scope of the `foo` function. If the user tries to access `x` outside of it,
they will get an error. So, I imagine that this means to solve `foo(x)` I will have to create a copy of
the global dictionary but with an update - the copy will hold an expression for `x`, created just before.
That copy will then be used by a (new) parser to create an expression for a given particular use of `foo(x)` -
`foo(1)`, `foo(3*4)`, `foo(a+b)`, etc., and within the scope of the function the parser will be able to parse
expressions using `x`. But the moment we jump out of `foo`, we will get back to using the global dictionary,
which has no idea what `x` is.

This PR introduces the ability to cope the dictionary and provide updates while doing it. It also copies
the special values counter so that new special values, created during parsing of the function, will not get
mixed with the higher-order ones.