feat: Different rules in preprocessor for left and right sides of an assignment
Currently in the preprocessor I have logic for removing whitespaces and for turning parentheses into special values.
This is going to get much more complicated with introduction of functions, since functions also use parentheses,
but in a different way. This PR anticipates that a little bit - if the line is an assignment, removing parentheses
will be now done only on the right side of it. The left side will stay the same and will be handled in the parser.