feat: When looking for parentheses in the preprocessor, ignore function parentheses
I use the same characters for denoting standard mathematical parentheses, like in `(1+2)*3`
and in functions, but function parentheses work differently. In the preprocessor, where the program
looks for pairs of parentheses to replace them with values, it should recognize when the opening
parenthesis is one of a function and it should ignore it. However, it should not ignore mathematical
parentheses in the place of arguments. For example: `1 + foo(x, y) + 2` should be left with no changes,
but `1 + foo((2 + 3) * 4, 5 / (6 - 7)) + 8` should have `(2 + 3)` and `(6 - 7)` replaced with values.