feat: Preprocessor flags
Introduces flags for each step of the preprocessor. The default configuration is to have all of them set to true.
Disabling some of the flags may help with unit tests. Also, in the future, there might be special cases when
I want to use only a part of the preprocessor functionality.