feat: Put parentheses around functions arguments
https://github.com/makingthematrix/replcalc/projects/1#card-74102362

This one is a bit hacky. In preparation for implementing functions I thought about a potential problem I can run into.
Consider something like:
`foo(1, bar(2, 3), 4)`
We have a function `foo` which has three arguments: `1`, `bar(2, 3)`, and `3`.
But how do I parse those arguments? Well, by looking for commas. So let's try to split the arguments of `foo` by commas.
`1`, `bar(2`, `3)`, `4`
The problem is, I need to ignore that one comma which belongs to the `bar` function, but it's not easy to do.
I would need to identify the `bar` function and jump over it. I can do it, I have the code for that, but in a different
place. Using it at this moment would be very messy.
Instead, I decided to make a new task in the preprocessor which puts parentheses around function arguments.
So `foo(1, bar(2, 3), 4)` becomes `foo(1, (bar(2, 3)), 4)` (constant arguments don't need parentheses).
And then in the next task, the one about removing parentheses, this becomes `foo(1, $1, 4)` because the argument with
parentheses is turned into a special value. Together the two tasks simply the function expression so it's pretty easy
to parse it - no need to worry about nested expressions in arguments.