refactor: Put logic about finding matching parentheses in one place
Looking for parentheses forces me to deal with some complexity I don't really like.
It is possible that there are no parentheses in the line and then the code should return 'None'
and this is fine, the program should move on. Or there can be parentheses and then I need to find
them and do some operations on the part of the line before the opening one, some on what's after
the closing one and yet something (probably the most interesting part) on what's between them.
But then, it is also possible that there is the opening parenthesis but no closing one and then
the code should return an error (if there is only the closing parenthesis it will also return
an error but in another place).

So, I'm mostly interested in the case where I have the parentheses but I still need to write code
for the two other cases.

This PR takes care of that with the help of a pattern I use sometimes in Scala:
I make a method `withParens(line){ body }` that takes a function (a body) which will be called only
if there are parentheses. If there are none, or if thre's only one, the method itself will handle that.

This will be useful in the next PRs where I start to implement functions.