feat: The Variable expression type
https://github.com/makingthematrix/replcalc/projects/1#card-74638105

So in the end it turned out that there is no need for a new expression types for function arguments.
Arguments can be variables (and the Variable expression type is the rewritten Value expression type).
The trick is in separating creation from usage. We have expression types Assignment and FunctionAssignment
for creating new variables and functions respectively when there is the "=" sign in the expression.
In that case we assume that what's on the left side of "=" is the name of the new variable or function
(and eventually the list of arguments) and what's on the right side is an expression we either evaluate
to a constant, in case of a variable, or parse it but not evaluate it in case of a function. And then
we put this new variable or function in the dictionary.

Later, when we parse an expression, we may find a name which we recognize as a name of a variable of a function
from the dictionary. If so, we get the associated expression and evaluate it. It's very straightforward
in the case of a variable, because then the original expression is already evaluated and now we have a constant,
but in the case of a function we evaluate it only at this point.

This distinction is important because it means that if we reassign a variable (this is now possible) then
it will not affect expressions which already used that variable. In those expressions the variable was
already evaluated and now we have a constant instead of it. Only in new expressions the new value of the variable
will be used. But if a variable is used in a function, that function will ask for its value at the point
of its own evaluation - so if the variable is reassigned, the function will get the new value.

An example:
```
x = 1
y = x + 1
z() = x + 1
y // returns: 2
z() // returns: 2
x = 3
y // it's a variable so it still returns 2
z() // it's a function so now it returns 4
```

By the way, the syntax makes it possible to drop parentheses in a call to a function without arguments, so both `z()`
and `z` are valid.

It's not possible to reassign a function. `z = 2` or `z() = 2` won't work.