refactor: Cleaner error handling
So far error handling was done "by hand", so to say. There is a lot of places in the code where I parse or evaluate
an expression and need to handle different outcomes. The result of parsing might be that the parsed expression is not
of the type I assumed - which is not exactly an error, I just need to try with another expression type. But it also
might be of the given type but it's incorrect - then it's really an error. Or, at last, it can be a valid expression
of the given type and then I need to handle the parsed entity in a certain one.

Until now all those cases were handled next to each other which made the code less readable than it could be.
This PR tries to move handling of errors and the "not this type" answer to the side and focus more on the "happy path"
when there is a valid expression.
I tried a few solutions and in the end decided on sticking to `type ParsedExpr[T] = Option[Either[Error, T]]` where
`T` is an expression type, but I introduced a few utility methods that take care of cases where I get `None` or
`Some(Left(error))` as results of parsing. The final goal is readability.