refactor: Code beautification
A bunch of cosmetic changes and small simplifications. Logic entities (the parser and the dictionary) passed around in
arguments were moved to first places in the arguments lists. Some logic was moved around to make code more readable.
Variable names were standarized (although I still use interchangeably `dict` and `dictionary`, and `expr` and `expression`.
`if / then / else` was changed to `match / case` in a few places.