refactor: Decouple Preprocessor
Not much value in this PR except maybe for easier creation of a preprocessor for preprocessor unit tests.
The real reason to do it was that I see the preprocessor and the parser as equal components talking to
each other, while so far the parser owned the preprocessor. Now they can be created separately, even though
they still need each other to work.