fix: Clean special assignments
Every time the user types in an expression with parentheses, special assignments are put in the dictionary.
They are used to process the expression but they are not used later. In practice they could just stay in
the dictionary forever - the user would never manage to use up all the memory - but it also doesn't hurt
to clean them after the expression is processed.

I also used this opportunity to fix a bug where numbers put directly after the closing parenthesis could be
interpreted as a part of the special assignment's name (e.g. "(1+2)3" would turn into "$13" where it should
be "$1" and "3" should be recognized as an error).