feat: Handle floating point imprecision
A final touch. Sometimes, with more complex equations, you may find out that zero is not always zero,
due to floating point issues on JVM (and basically everywhere else). I tried to remedy it with
calculation a magic number that will be used as a precision point for expressions: any result closer
to zero than this will be considered to be zero.
Therese is a bit of boilerplate code but it allows for adding more checks like this in the future.