refactor: Additional preprocessor decoupling
After giving it some thought I decided to treat `Preprocessor` the same as `Parser`.
Now it can also be created without the other one, and then the new `setup` method can be used to connect them.
There is an utility `apply` method in `Parser` to do it, so `Parser` can be still seen as a bit more
important of the two.

Plus a few minor changes to `Dictionary`. From now on a valid variable name can't consist of only "_".