refactor: Parse function line
A part of line of the form `functionName(arg1,arg2,...)` is now handled in two places - it can be a function assignment,
or a function evaluated in an expression. On top of that, I'm working on a fix to a bug that will require to handle it
once more, in the preprocessor. To make it easier, I decided to extract the logic that turns such a piece of text into
a small case class to a separate function.