fix: Function arguments used inside parentheses
A very nasty bug. Two in fact, but connected. It involves function arguments and parentheses replaced by special variables.

Let's consider a function assignment which looks like this: `f(x) = (1 + x)`. There would be no problem if no parentheses
were used. But the way it is now, the preprocessor tries to put the nested expression `1+x` in the dictionary under
a special name and then replace the parentheses with it. The problem is, to do it, `1+x` must be parsed, and the parser
doesn't know that `x` is a function argument. It would know if there were no parentheses because then the whole line
would be parsed as a function assignment, but here the parser has only the chunk `1+x` and no information that this is
a function assignment.

To fix it, I make a copy of the original parser, updated with the function arguments. This way parsing works now, but
it creates additional complexity: all special variables made when removing parentheses are now put into the new parser,
not the original one. So after it's done, I need to update the original one with new special variables from the new one.

And that led me to the second bug: Some time ago I introduced cleaning the dictionary from special variables, thinking that
after an expression is evaluated, the special variables used in evaluation are no longer needed. But that's not true
for function assignments. They are evaluated on demand, possibly many times, and they require the special variables to
still be in the dictionary. So cleaning would have to be much smarter than just removing special variables - it would have to
look through function assignments and figure out which special variables are in use.
Instead, I removed cleaning for now. I assume that since this is a simple REPL calculator, then there is no risk in
keeping the special variables in the memory. There will never be too many of them.

There is a way to fix both these bugs in a more scalable way: A function should carry its own dictionary with arguments,
special variables, and everything else needed to evaluate it. It's called "a stack frame" and it's how compilers deal
with functions (as far as I remember from my studies). But since this is just a calculator, I decided to stop at this point.
But if I was to develop this project further, I would certainly consider to rewrite this functionality like that.