refactor: Move ParsedFunction to a separate file
ParsedFunction is a small utility class for help in parsing a function or a function assignment out of a text chunk.
If the text is of the form `name(arg1, arg2, ...)`, the parsed function will keep the name and the list of
arguments. I will also properly return a `None` if the text doesn't seem to be a function (i.e. it does not have
parentheses) and it will return an error if it seems to be a function, but it can't be parsed.