fix: An infinite recursion in evaluation when an argument has the same name as a variable
If the user first defines a variable, then defines a function with an argument name the same as the variable,
and then use the variable in place of the argument to call the function, the program gets confused.
Consider this:
x = 1
f(x) = x + 2
f(x)
The last line is a call to the function f with the argument being *the variable* x, so it should be f(1),
and the result should be 3. The bug was that I first updated the dictionary with the argument x, replacing
the variable x, and only then tried to evaluate the variable x. Now, instead of the constant 1, the program
found in the dictionary the variable x (arguments are added to the dictionary as variables), tried to evaluate
it, and by evaluating a variable it again looked in the dictionary for x, found the variable x... and so on.

The fix is to first evaluate arguments to constants, and only then update the dictionary with the arguments.